% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brca.data-package.r
\docType{package}
\name{brca.data}
\alias{brca.data}
\alias{brca.data-package}
\title{brca.data: Breast Cancer data from TCGA}
\description{
To use this package load the following datasets:
}
\details{
- `tissue.ix`: indexes of all tissues;
- `tissue`: gene expression data for all types of tissues, see `names(tissue)`;
- `tissue.barcode`: Patient's participation data code (TCGA-XX-XXXX) per type of tissue;
- `clinical`: Clinical data per tissue type. Has the same structure as tissue.

To see how the data was retrieved, check the `brca_import_tcga.Rmd`
}

