% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_assay.R
\name{build.assay}
\alias{build.assay}
\title{Create MultiAssayExperiment object from data}
\usage{
build.assay(clinical.custom = NULL, gdc.custom = NULL,
  mutation.custom = NULL, rnaseq.custom = NULL)
}
\arguments{
\item{clinical}{use custom clinical (that can be pre-processed)}
}
\value{
a MultiAssayExperiment object
}
\description{
Create MultiAssayExperiment object from data
}
\examples{
assay <- build.assay()
assay[['RNASeq']]
assar$vital_status
}
