% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{clinical}
\alias{clinical}
\title{Clinical data per tissue type}
\format{An object of class \code{list} of length 5.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
clinical
}
\description{
A dataset containing a list of dataframes with clinical data per tissue type.
}
\keyword{datasets}
