% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fpkm.per.tissue.barcode}
\alias{fpkm.per.tissue.barcode}
\title{Patient's barcode per tissue type}
\format{An object of class \code{list} of length 5.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
fpkm.per.tissue.barcode
}
\description{
A dataset containing a list of vectors with TCGA barcodes per tissue type.
}
\keyword{datasets}
