% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fpkm.per.tissue}
\alias{fpkm.per.tissue}
\title{Gene expression levels per tissue type.}
\format{A list with FPKM gene expression data.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
fpkm.per.tissue
}
\description{
List of matrices with RNASeq v2 FPKM (rows: genomic ranges) x (columns: patients)
}
\details{
FPKM: fragments per kilobase per million
= [# of fragments]/[length of transcript in kilo base]/[million mapped reads]

For all FPKM expression data matrix, run `joinRNASeqData` dataset
}
\keyword{datasets}
