% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gdc.original}
\alias{gdc.original}
\title{Original data from GDC}
\format{An object of class \code{list} of length 16.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
gdc.original
}
\description{
Contains:
- All clinical data (clinical, follow-up, etc..)
- Gene expression metadata (FPKM and Counts)
- Mutation data
}
\keyword{datasets}
