% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gdc}
\alias{gdc}
\title{Original data from GDC}
\format{An object of class \code{list} of length 8.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
gdc
}
\description{
Contains:
- All clinical data (clinical, follow-up, etc..)
- Gene expression data
- Mutation data
}
\keyword{datasets}

