% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gene.ranges}
\alias{gene.ranges}
\title{Genomic ranges with description for genes in tissue[.all] matrices}
\format{A list with RPKM gene expression data.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
gene.ranges
}
\description{
Contains ensembl id and external gene names.
}
\keyword{datasets}

