% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getParticipantCode}
\alias{getParticipantCode}
\title{Get Participant part of TCGA Barcode}
\usage{
getParticipantCode(fullBarcode)
}
\arguments{
\item{fullBarcode}{string, or  vector of strings}
}
\value{
participant's barcode
}
\description{
TCGA-XX-XXXX
}
\examples{
getParticipantCode('TCGA-3C-AAAU-01A-11R-A41B-07')
getParticipantCode(c('TCGA-3C-AAAU-01A-11R-A41B-07', 'TCGA-3C-AALI-01A-11R-A41B-07'))
}
