% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSampleTypeCode}
\alias{getSampleTypeCode}
\title{Get Sample Type part of TCGA Barcode}
\usage{
getSampleTypeCode(fullBarcode)
}
\arguments{
\item{fullBarcode}{number, or  vector of numbers}
}
\value{
sample code
}
\description{
XX - where 01 is solid tumor, 06 is metasteses, etc..
}
\details{
See [Code Tables Report](https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/sample-type-codes) for a complete list of sample codes.
}
\examples{
getParticipantCode('TCGA-3C-AAAU-01A-11R-A41B-07')
getParticipantCode(c('TCGA-3C-AAAU-01A-11R-A41B-07', 'TCGA-3C-AALI-01A-11R-A41B-07'))
}
