% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{joinRNASeqData}
\alias{joinRNASeqData}
\title{Joins all expression data in a single matrix}
\usage{
joinRNASeqData()
}
\value{
a matrix with gene expression levels and all tissue samples
}
\description{
This is not stored directly as it would be duplicate information
 which takes a lot of storage
}
