% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loadGDCRnaSeq}
\alias{loadGDCRnaSeq}
\title{Get rnaseq assay data to GDC object}
\usage{
loadGDCRnaSeq()
}
\value{
a full gdc
}
\description{
This cannot be cached in the package as it takes too much space and
 would be redundant with fpkm.per.tissue data
}
