% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mutation}
\alias{mutation}
\title{List with mutation data}
\format{An object of class \code{list} of length 3.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
mutation
}
\description{
Contains:
- Count of mutations per each pair of gene/case
- Original data from gdc (filtered out duplicate rows)
- Duplicated rows that were filtered
}
\keyword{datasets}
