% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tissue.all}
\alias{tissue.all}
\title{Gene expression for all tissue samples.}
\format{A matrix with FPKM gene expression data.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
tissue.all
}
\description{
Matrix with all samples gene expression from RNASeq v2 FPKM (rows: genomic ranges) x (columns: patients)
}
\details{
FPKM: fragments per kilobase per million
= [# of fragments]/[length of transcript in kilo base]/[million mapped reads]
}
\keyword{datasets}

