% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tissue.ix}
\alias{tissue.ix}
\title{Index of patients per tissue.}
\format{An object of class \code{list} of length 22.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
tissue.ix
}
\description{
A dataset containing a list of index tissue names from TCGA breast cancer dataset (BRCA).
}
\details{
Each element has a vector of logical index for the tissue data.
}
\keyword{datasets}

