% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tissue}
\alias{tissue}
\title{Gene expression per tissue.}
\format{A list with RPKM gene expression data.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
tissue
}
\description{
A dataset containing a list of tissue names from TCGA breast cancer dataset (BRCA).
}
\details{
Each element has a normalized gene expression (RPKM) matrix with genes (rows) x patients (columns).

For all gene expression data matrix, see `tissue.all` dataset
}
\keyword{datasets}

