% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gdc}
\alias{gdc}
\title{Original data from GDC}
\format{An object of class \code{list} of length 17.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
gdc
}
\description{
To get RNA-Seq assay data, please run loadGDCRnaSeq()
}
\details{
It is not included in the data originaly to avoid redundant
 data in the package.

Contains:
- All clinical data (clinical, follow-up, etc..)
- Gene expression metadata (assay can be obtained by function 'loadGDCRnaSeq')
- Mutation data
}
\keyword{datasets}
