% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gene.ranges}
\alias{gene.ranges}
\title{Genomic ranges with description for genes in tissue matrices}
\format{An object of class \code{GRanges} of length 57035.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
gene.ranges
}
\description{
Contains ensembl id and external gene names.
}
\keyword{datasets}
