% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tissue}
\alias{tissue}
\title{Gene expression per tissue.}
\format{A list with FPKM gene expression data.}
\source{
\url{https://gdc-portal.nci.nih.gov/projects/TCGA-BRCA}
}
\usage{
tissue
}
\description{
List of matrices with RNASeq v2 FPKM (rows: genomic ranges) x (columns: patients)
}
\details{
FPKM: fragments per kilobase per million
= [# of fragments]/[length of transcript in kilo base]/[million mapped reads]

For all gene expression data matrix, see `tissue.all` dataset (stored in a different variable for memory optimization)
}
\keyword{datasets}

